refactor: car test
